Modifications to allow for problems fetching the media from a new item.  If we get a new item from the feed and it doesn't match our regexp, stuff it in last_show immediately and stop looking.  If the new item does match, try to fetch the media, and if the http GET for the media isn't successful, don't update last_show and go through another sleep cycle before trying again. The last_show table won't get updated until we successfully download, normalize, and upload a new show to the webserver.
Also, rather than caching the rotated logfiles on the run server and pushing them repeatedly to the webserver, connect to the webserver via IPC::PerlSSH and rename them on the webserver.  Then we only need to copy the latest log file up to the webserver.  This can save several minutes of run time.